+++ src/Services/Catalog/DocumentContractor/Result/DocumentContractorItemResult.php
@@ -1,0 +1,26 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Dmitriy Ignatenko <algonexys@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Catalog\DocumentContractor\Result;
+
+use Bitrix24\SDK\Core\Result\AbstractAnnotatedItem;
+
+/**
+ * @property-read int $id
+ * @property-read int $documentId
+ * @property-read int $entityTypeId
+ * @property-read int $entityId
+ */
+class DocumentContractorItemResult extends AbstractAnnotatedItem
+{
+}